APEX-7921 Remove magewire breadcheckout component
