+++ view/frontend/templates/hyva/checkout/payment/method/breadcheckout.phtml
@@ -2,27 +2,25 @@
 /**
  * Bread BreadCheckout - Hyva Checkout Payment Method Template
  *
- * @var \Magento\Framework\View\Element\Template $block
+ * @var \Bread\BreadCheckout\Block\Hyva\Checkout\Payment\Method\BreadCheckout $block
  * @var \Magento\Framework\Escaper $escaper
- * @var \Bread\BreadCheckout\Magewire\Checkout\Payment\Method\BreadCheckout $magewire
  */
 
-$magewire = $block->getMagewireComponent();
-if (!$magewire) {
+if (!$block->isAvailable()) {
     return;
 }
 
-$integrationKey = $magewire->getIntegrationKey();
-$sdkName = $magewire->getSdkName();
-$jsLocation = $magewire->getJsLocation();
-$currencyCode = $magewire->getCurrencyCode();
-$checkoutData = $magewire->getCheckoutData();
-$billingContact = $magewire->getBillingContact();
-$shippingContact = $magewire->getShippingContact();
-$transactionId = $magewire->getTransactionId();
-$isApproved = $magewire->isApproved;
-$paymentUrl = $magewire->getPaymentUrl();
-$validateTotalsUrl = $magewire->getValidateTotalsUrl();
+$integrationKey = $block->getIntegrationKey();
+$sdkName = $block->getSdkName();
+$jsLocation = $block->getJsLocation();
+$currencyCode = $block->getCurrencyCode();
+$checkoutData = $block->getCheckoutData();
+$billingContact = $block->getBillingContact();
+$shippingContact = $block->getShippingContact();
+$transactionId = $block->getTransactionId();
+$isApproved = $block->isApproved();
+$paymentUrl = $block->getPaymentUrl();
+$validateTotalsUrl = $block->getValidateTotalsUrl();
 
 $breadConfig = [
     'integrationKey' => $integrationKey,
